many small changes
shifted fcn_fn to main file, added get_config, cleared 3 tests, fixed linting issues and docstrings, moved around conv layers to reduce output-shape computing calls
