+++ keras_cv/models/segmentation/fcn.py
@@ -1,53 +1,456 @@
-# Copyright 2022 The KerasCV Authors
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     https://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
 import tensorflow as tf
 
-from keras_cv.layers.preprocessing import Resizing
+from keras_cv.layers.preprocessing import Rescaling
 from keras_cv.models import VGG16
 from keras_cv.models import VGG19
-import numpy as np
 
 BACKBONE_CONFIG = {
-    "vgg16": {"BLOCK3": 11, "BLOCK4": 15, "BLOCK5": 19, "DENSE_START": -3},
-    "vgg19": {"BLOCK3": 12, "BLOCK4": 17, "BLOCK5": 22, "DENSE_START": -3},
+    "vgg16": {
+        "BLOCK3": "block3_pool",
+        "BLOCK4": "block4_pool",
+        "BLOCK5": "block5_pool",
+        "DENSE_START": "fc1",
+        "DENSE_END": "predictions",
+    },
+    "vgg19": {
+        "BLOCK3": "block3_pool",
+        "BLOCK4": "block4_pool",
+        "BLOCK5": "block5_pool",
+        "DENSE_START": "fc1",
+        "DENSE_END": "predictions",
+    },
 }
+BACKBONE = {"vgg16": VGG16, "vgg19": VGG19}
 
 
-class BilinearInitializer(tf.keras.initializers.Initializer):
-    """Initializer for Conv2DTranspose to perform bilinear interpolation on each channel."""
+def getPoolLayers(model, pool_id):
+    """Function to get the all layers upto a certain named layer. Used only for extracting MaxPooling2D-based layers.
 
-    def __call__(self, shape, dtype=None, **kwargs):
-        kernel_size, _, filters, _ = shape
-        arr = np.zeros((kernel_size, kernel_size, filters, filters))
-        ## make filter that performs bilinear interpolation through Conv2DTranspose
-        upscale_factor = (kernel_size + 1) // 2
-        if kernel_size % 2 == 1:
-            center = upscale_factor - 1
+    Args:
+        model: `tf.keras.layers.Model` or `tf.keras.Sequential`, The original model from which the layers are extracted
+        pool_id: str, The name of the layer that acts as the end layer limit for the extraction
+    Returns:
+        layers: `tf.keras.Sequential` representing the extracted layers
+    """
+    layer_names = [layer.name for layer in model.layers]
+
+    pool_layers = []
+    for layer_id in layer_names:
+        layer = model.get_layer(layer_id)
+        pool_layers.append(layer)
+        if layer_id == pool_id:
+            return tf.keras.Sequential(pool_layers)
+
+
+def getDenseToConvolutionLayers(model, dense_start_id, dense_end_id):
+    """Utility function to convert all `tf.keras.layers.Dense` functions to `tf.keras.layers.Conv2D` with `filters=units` from the Dense layers and `kernel_size=(1,1)`, `padding='same'`
+
+    Args:
+        model: `tf.keras.layers.Model` or `tf.keras.Sequential`, The original model from which the layers are extracted
+        dense_start_id: str, The name of the layer that acts as the start layer limit for the extraction
+        dense_end_id: str, The name of the layer that acts as the end layer limit for the extraction
+    Returns:
+        layers: `tf.keras.Sequential` representing the converted layers.
+    """
+    layer_names = [layer.name for layer in model.layers]
+    dense_flag = False
+
+    units = []
+    dense_convs = []
+
+    for layer_id in layer_names:
+        if layer_id == dense_start_id or dense_flag is True:
+            dense_flag = True
+            layer = model.get_layer(layer_id)
+            units.append(layer.units)
+            if layer_id == dense_end_id:
+                dense_flag = False
+                break
+
+    for filter_idx in range(len(units)):
+        filter = units[filter_idx]
+        if filter_idx == 0:
+            dense_layer = tf.keras.layers.Conv2D(
+                filters=filter,
+                kernel_size=(
+                    7,
+                    7,
+                ),
+                strides=(1, 1),
+                activation="relu",
+                padding="same",
+            )
+            dense_convs.append(dense_layer)
         else:
-            center = upscale_factor - 0.5
-        og = np.ogrid[:kernel_size, :kernel_size]
-        kernel = (1 - np.abs(og[0] - center) / upscale_factor) * (
-            1 - np.abs(og[1] - center) / upscale_factor
-        )  # kernel shape is (kernel_size, kernel_size)
-        for i in range(filters):
-            arr[..., i, i] = kernel
-        return tf.convert_to_tensor(arr, dtype=dtype)
+            dense_layer = tf.keras.layers.Conv2D(
+                filters=filter,
+                kernel_size=(1, 1),
+                strides=(1, 1),
+                activation="relu",
+                padding="same",
+            )
+            dense_convs.append(dense_layer)
+        dropout_layer = tf.keras.layers.Dropout(0.5)
+        dense_convs.append(dropout_layer)
+    return tf.keras.Sequential(dense_convs)
+
+
+def VGGBackboneBuilder(
+    model_version,
+    model_architecture,
+    include_rescaling,
+    classes,
+    input_shape=(224, 224, 3),
+    include_imagenet_weights=False,
+):
+    """Utility function to build the backbone of the FCN with variants of the VGG model.
+
+    Args:
+        model_version: str, one of 'vgg16' or 'vgg19'. Defines which model to use as the backbone base.
+        model_architecture: str, one of 'fcn8s', 'fcn16s' or 'fcn32s'.
+            Defines which architecture and sampling method should be used as detailed in the
+            paper [Fully Convolutional Networks for Semantic Segmentation](https://arxiv.org/pdf/1411.4038.pdf)
+        include_rescaling: bool, one of True or False. Defines whether to use a Rescaling layer or not.
+        classes: int. Defines the number of classes for the output.
+        input_shape: `list` or `tuple`. Defines the shape of the tensor to be expected as input.
+
+    Returns:
+        model: tf.keras.models.Model. Represents the graph of all backbone operations and outputs for the chosen architecture.
+    """
+    if include_imagenet_weights:
+        weights = "imagenet"
+    else:
+        weights = None
+
+    vgg_model = BACKBONE[model_version](
+        include_rescaling=include_rescaling,
+        classes=classes,
+        input_shape=input_shape,
+        include_top=True,
+        weights=weights,
+    )
+
+    x = tf.keras.Input(shape=input_shape)
+
+    if model_architecture == "fcn8s":
+        # Made it like this, because then parameter sharing occurs to get the model parameter size to go down drastically
+        pool5 = getPoolLayers(vgg_model, BACKBONE_CONFIG[model_version]["BLOCK5"])
+
+        pool3_output, pool4_output = x, x
+
+        for layer in pool5.layers:
+            pool3_output = layer(pool3_output)
+            if layer.name == BACKBONE_CONFIG[model_version]["BLOCK3"]:
+                break
+        for layer in pool5.layers:
+            pool4_output = layer(pool4_output)
+            if layer.name == BACKBONE_CONFIG[model_version]["BLOCK4"]:
+                break
+
+        dense_convs = getDenseToConvolutionLayers(
+            model=vgg_model,
+            dense_start_id=BACKBONE_CONFIG[model_version]["DENSE_START"],
+            dense_end_id=BACKBONE_CONFIG[model_version]["DENSE_END"],
+        )
+        pool5_output = pool5(x)
+        pool5_output = dense_convs(pool5_output)
+
+        model = tf.keras.models.Model(
+            inputs=x,
+            outputs={
+                "pool3": pool3_output,
+                "pool4": pool4_output,
+                "pool5": pool5_output,
+            },
+        )
+
+        return model
+    elif model_architecture == "fcn16s":
+
+        pool5 = getPoolLayers(vgg_model, BACKBONE_CONFIG[model_version]["BLOCK5"])
+
+        pool4_output = x
+
+        for layer in pool5.layers:
+            pool4_output = layer(pool4_output)
+            if layer.name == BACKBONE_CONFIG[model_version]["BLOCK4"]:
+                break
+
+        dense_convs = getDenseToConvolutionLayers(
+            model=vgg_model,
+            dense_start_id=BACKBONE_CONFIG[model_version]["DENSE_START"],
+            dense_end_id=BACKBONE_CONFIG[model_version]["DENSE_END"],
+        )
+        pool5_output = pool5(x)
+        pool5_output = dense_convs(pool5_output)
+
+        model = tf.keras.models.Model(
+            inputs=x, outputs={"pool4": pool4_output, "pool5": pool5_output}
+        )
+
+        return model
+    elif model_architecture == "fcn32s":
+
+        pool5 = getPoolLayers(vgg_model, BACKBONE_CONFIG[model_version]["BLOCK5"])
+
+        dense_convs = getDenseToConvolutionLayers(
+            model=vgg_model,
+            dense_start_id=BACKBONE_CONFIG[model_version]["DENSE_START"],
+            dense_end_id=BACKBONE_CONFIG[model_version]["DENSE_END"],
+        )
+        pool5_output = pool5(x)
+        pool5_output = dense_convs(pool5_output)
+
+        model = tf.keras.models.Model(inputs=x, outputs={"pool5": pool5_output})
+
+        return model
+
+
+def VGGArchitectureBuilder(
+    model_version,
+    model_architecture,
+    include_rescaling,
+    classes,
+    input_tensor,
+    input_shape=(224, 224, 3),
+    include_imagenet_weights=False,
+):
+    """Main function for development and execution of full FCN model specifically for VGG-based backbones.
+
+    Args:
+        model_version: str, one of 'vgg16' or 'vgg19'. Defines which model to use as the backbone base.
+        model_architecture: str, one of 'fcn8s', 'fcn16s' or 'fcn32s'.
+            Defines which architecture and sampling method should be used as detailed in the
+            paper [Fully Convolutional Networks for Semantic Segmentation](https://arxiv.org/pdf/1411.4038.pdf)
+        include_rescaling: bool, one of True or False. Defines whether to use a Rescaling layer or not.
+        classes: int. Defines the number of classes for the output.
+        input_shape: `list` or `tuple`. Defines the shape of the tensor to be expected as input.
+    Returns:
+        model: `tf.Tensor` representing the output of the FCN model. Shape should match the
+        format (input_shape[1], input_shape[2], classes)
+
+    """
+    if model_version not in ["vgg16", "vgg19"]:
+        raise ValueError(
+            "Chosen `backbone` argument is not a valid allowed backbone. Possible options are ['vgg16', 'vgg19']"
+        )
+    if model_architecture == "fcn8s":
+        backbone = VGGBackboneBuilder(
+            model_version=model_version,
+            include_rescaling=include_rescaling,
+            classes=classes,
+            input_shape=input_shape,
+            model_architecture="fcn8s",
+            include_imagenet_weights=include_imagenet_weights,
+        )
+        backbone_output = backbone(input_tensor)
+        pool3, pool4, pool5 = (
+            backbone_output["pool3"],
+            backbone_output["pool4"],
+            backbone_output["pool5"],
+        )
+
+        pool3 = tf.keras.layers.Conv2D(
+            filters=classes,
+            kernel_size=(1, 1),
+            padding="same",
+            strides=(1, 1),
+            activation="relu",
+        )(pool3)
+
+        pool4 = tf.keras.layers.Conv2D(
+            filters=classes,
+            kernel_size=(1, 1),
+            padding="same",
+            strides=(1, 1),
+            activation="relu",
+        )(pool4)
+
+        pool5 = tf.keras.layers.Conv2D(
+            filters=classes,
+            kernel_size=(1, 1),
+            padding="same",
+            strides=(1, 1),
+            activation="relu",
+        )(pool5)
+
+        pool5_upsampling = tf.keras.layers.UpSampling2D(
+            size=(2, 2), data_format="channels_last", interpolation="bilinear"
+        )
+        pool5 = pool5_upsampling(pool5)
+
+        intermediate_pool_output = tf.keras.layers.Add()([pool4, pool5])
+
+        intermediate_pool_output = tf.keras.layers.UpSampling2D(
+            size=(2, 2), data_format="channels_last", interpolation="bilinear"
+        )(intermediate_pool_output)
+
+        final_pool_output = tf.keras.layers.Add()([pool3, intermediate_pool_output])
+
+        output_conv_layer = tf.keras.layers.Conv2D(
+            filters=classes,
+            kernel_size=(1, 1),
+            activation="softmax",
+            padding="same",
+            strides=(1, 1),
+        )
+
+        output_upsample_layer = tf.keras.layers.UpSampling2D(
+            size=(8, 8), data_format="channels_last", interpolation="bilinear"
+        )
+
+        final_output = output_conv_layer(final_pool_output)
+        final_output = output_upsample_layer(final_output)
+
+        return final_output
+
+    elif model_architecture == "fcn16s":
+        backbone = VGGBackboneBuilder(
+            model_version=model_version,
+            include_rescaling=include_rescaling,
+            classes=classes,
+            input_shape=input_shape,
+            model_architecture="fcn16s",
+            include_imagenet_weights=include_imagenet_weights,
+        )
+        backbone_output = backbone(input_tensor)
+        pool4, pool5 = backbone_output["pool4"], backbone_output["pool5"]
+
+        pool4 = tf.keras.layers.Conv2D(
+            filters=classes,
+            kernel_size=(1, 1),
+            padding="same",
+            strides=(1, 1),
+            activation="relu",
+        )(pool4)
+
+        pool5 = tf.keras.layers.Conv2D(
+            filters=classes,
+            kernel_size=(1, 1),
+            padding="same",
+            strides=(1, 1),
+            activation="relu",
+        )(pool5)
+
+        pool5_upsampling = tf.keras.layers.UpSampling2D(
+            size=(2, 2), data_format="channels_last", interpolation="bilinear"
+        )
+        pool5 = pool5_upsampling(pool5)
+
+        final_pool_output = tf.keras.layers.Add()([pool4, pool5])
+
+        output_conv_layer = tf.keras.layers.Conv2D(
+            filters=classes,
+            kernel_size=(1, 1),
+            activation="softmax",
+            padding="same",
+            strides=(1, 1),
+        )
+
+        output_upsample_layer = tf.keras.layers.UpSampling2D(
+            size=(16, 16), data_format="channels_last", interpolation="bilinear"
+        )
+
+        final_output = output_conv_layer(final_pool_output)
+        final_output = output_upsample_layer(final_output)
+
+        return final_output
+
+    elif model_architecture == "fcn32s":
+        backbone = VGGBackboneBuilder(
+            model_version=model_version,
+            include_rescaling=include_rescaling,
+            classes=classes,
+            input_shape=input_shape,
+            model_architecture="fcn32s",
+            include_imagenet_weights=include_imagenet_weights,
+        )
+        backbone_output = backbone(input_tensor)
+        pool5 = backbone_output["pool5"]
+
+        pool5 = tf.keras.layers.Conv2D(
+            filters=classes,
+            kernel_size=(1, 1),
+            padding="same",
+            strides=(1, 1),
+            activation="relu",
+        )(pool5)
+
+        pool5_upsampling = tf.keras.layers.UpSampling2D(
+            size=(32, 32), data_format="channels_last", interpolation="bilinear"
+        )
+
+        output_conv_layer = tf.keras.layers.Conv2D(
+            filters=classes,
+            kernel_size=(1, 1),
+            activation="softmax",
+            padding="same",
+            strides=(1, 1),
+        )
+
+        final_output = output_conv_layer(pool5)
+        final_output = pool5_upsampling(final_output)
+
+        return final_output
+
+
+def CustomArchitectureBuilder(model):
+    """Utility function to parse a tf.keras.Model structure and get a FCN backbone from it.
+        It maps all Conv2D and MaxPooling layers directly as a 1-to-1 port, while it
+        converts all Dense layers into 1x1 Conv2D layers.
+
+    Args:
+        model: `tf.keras.models.Model` instance. Defines the custom backbone model passed as an argument.
+
+    Returns:
+        backbone: `tf.keras.Sequential` instance defining the FCN-specific parse output for the model passed as argument.
+
+    Warning:
+        It has undefined behaviour for ResNet-style, skip-connection based models.
+    """
+    layer_list = []
+    dense_units_list = []
+    # Design choice : Either fully reject, or pick up only Conv2D and MaxPooling2D layers
+    # and convert Dense to Conv in `build()`
+    # Possible edge case : ResNets
+    # Current design allows a simple CNN with Conv2D, MaxPooling2D and Dense layers only,
+    # to be parsed and made into a FCN directly.
+    for i in model.layers:
+        if isinstance(i, tf.keras.layers.Conv2D) or isinstance(
+            i, tf.keras.layers.MaxPooling2D
+        ):
+            layer_list.append(i)
+        elif isinstance(i, tf.keras.layers.Dense):
+            dense_units_list.append(i.units)
+            dense_convs = [
+                tf.keras.layers.Conv2D(
+                    filters=filters,
+                    kernel_size=(1, 1),
+                    strides=(1, 1),
+                    activation="relu",
+                    padding="same",
+                )
+                for filters in dense_units_list
+            ]
+            layer_list.append(*dense_convs)
+    if len(layer_list) == 0:
+        raise ValueError(
+            "Entered `backbone` argument does not have any Conv2D or MaxPooling layers. Include a `tf.keras.models.Model` with `keras.layers.Conv2D` or `keras.layers.MaxPooling2D` layers"
+        )
+    if not all(
+        isinstance(layer, tf.keras.layers.Conv2D)
+        or isinstance(layer, tf.keras.layers.MaxPooling2D)
+        or isinstance(layer, tf.keras.layers.Dense)
+        for layer in layer_list
+    ):
+        raise ValueError(
+            "Entered `backbone` argument has custom layers. Include a `tf.keras.models.Model` with `tf.keras.layers.Conv2D`, `tf.keras.layers.Dense` or `tf.keras.layers.MaxPooling2D` layers only."
+        )
+    return tf.keras.Sequential(layer_list)
 
 
 class FCN(tf.keras.models.Model):
-    """A segmentation model based on the Fully Convolutional Network introduced by Long et. al.
+    """A segmentation model based on the [Fully Convolutional Networks for Semantic Segmentation](https://arxiv.org/pdf/1411.4038.pdf) paper introduced by Long et. al.
 
     Args:
         classes: int, the number of classes for the segmentation model. Note that
@@ -69,309 +472,206 @@ class FCN(tf.keras.models.Model):
             2. 'fcn16s', a FCN-16S definition
             3. 'fcn32s', a FCN-32S definition
             Defaults to 'fcn8s'.
-        include_resizing: bool, defines whether to include a `tf.keras.layers.Resizing` layer which defaults to a height and width of value 224. Defaults to `False`
-
+        input_shape: `list` or `tuple`. Defines the shape of the tensor to be expected as input.
+        include_rescaling: bool, one of True or False. Defines whether to use a Rescaling layer or not.
+        return_mask: bool, one of True or False. Returns a 1-channel result instead of a
+            num_classes-channel result.
+        return_dtype: `tf.dtypes`. Return results in the chosen dtype. Defaults to tf.float32
+        include_imagenet_weights: bool, one of True or False. Defines whether to use
+            original ImageNet weights for VGG backbones.
     """
 
     def __init__(
         self,
         classes,
-        backbone="vgg16",
+        backbone,
         model_architecture=None,
-        include_resizing=False,
-        **kwargs
+        input_shape=(224, 224, 3),
+        include_rescaling=False,
+        return_mask=False,
+        return_dtype=tf.float32,
+        include_imagenet_weights=False,
     ):
-        super(FCN, self).__init__(**kwargs)
 
-        self.resize = include_resizing
-        self.num_classes = classes
-        self.backbone = backbone
-        self.model_architecture = model_architecture
-
-        if self.resize:
-            self.resizing_layer = Resizing(
-                height=224, width=224, interpolation="bilinear"
-            )
-
-    def build(self, input_shape):
-        self.height = input_shape[1]
-        self.width = input_shape[2]
-
-        if isinstance(self.backbone, str):
-
-            if self.backbone == "vgg16":
-                # Assumes `channels_last` format
-                if len(input_shape) == 3:
-                    self.vgg_input_shape = (
-                        input_shape[0],
-                        input_shape[1],
-                        input_shape[2],
-                    )
-                elif len(input_shape) == 4:
-                    self.vgg_input_shape = (
-                        input_shape[1],
-                        input_shape[2],
-                        input_shape[3],
-                    )
-                self.backbone_base_model = VGG16(
-                    include_rescaling=False,
-                    include_top=True,
-                    classes=self.num_classes,
-                    input_shape=self.vgg_input_shape,
-                )
-                self.backbone_name = self.backbone
-            elif self.backbone == "vgg19":
-                self.vgg_input_shape = (input_shape[1], input_shape[2], input_shape[3])
-                self.backbone_base_model = VGG19(
-                    include_rescaling=False,
-                    include_top=True,
-                    classes=self.num_classes,
-                    input_shape=self.vgg_input_shape,
-                )
-                self.backbone_name = self.backbone
-            else:
+        if isinstance(backbone, tf.keras.models.Model):
+            if model_architecture is not None:
                 raise ValueError(
-                    "Entered `backbone` argument is not a standard available backbone. Valid options are ['vgg16', 'vgg19']"
+                    "`model_architecture` cannot be set if `backbone` is not in ['vgg16', 'vgg19']. Either set `backbone` to one of the accepted values or remove the `model_architecture` argument."
                 )
-
-        else:
-            if isinstance(self.backbone, tf.keras.models.Model):
-                layer_list = []
-                # Design choice : Either fully reject, or pick up only Conv2D and MaxPooling2D layers and convert Dense to Conv in `build()`
-                # Possible edge case : ResNets
-                # Current design allows a simple CNN with Conv2D, MaxPooling2D and Dense layers only, to be parsed and made into a FCN directly.
-                for i in self.backbone.layers:
-                    if isinstance(i, tf.keras.layers.Conv2D) or isinstance(
-                        tf.keras.layers.MaxPooling2D
-                    ):
-                        layer_list.append(i)
-                if len(layer_list) == 0:
-                    raise ValueError(
-                        "Entered `backbone` argument does not have any Conv2D or MaxPooling layers. Include a `tf.keras.models.Model` with `keras.layers.Conv2D` or `keras.layers.MaxPooling2D` layers"
-                    )
-                if not all(
-                    isinstance(layer, tf.keras.layers.Conv2D)
-                    or isinstance(layer, tf.keras.layers.MaxPooling2D)
-                    or isinstance(layer, tf.keras.layers.Dense)
-                    for layer in layer_list
-                ):
-                    raise ValueError(
-                        "Entered `backbone` argument has custom layers. Include a `tf.keras.models.Model` with `tf.keras.layers.Conv2D` or `tf.keras.layers.MaxPooling2D` layers only."
-                    )
-                self.backbone = tf.keras.Sequential(layer_list)
-                self.backbone_base_model = self.backbone
-            else:
+            if include_imagenet_weights:
                 raise ValueError(
-                    "Unsupported type. Valid types are `tf.keras.models.Model`"
+                    """`include_imagenet_weights` cannot be set if `backbone` is not in ['vgg16', 'vgg19']. Either set `backbone` to one of the accepted values or remove the
+                    `include_imagenet_weights` argument."""
                 )
-
-        if self.model_architecture not in ["fcn8s", "fcn16s", "fcn32s"]:
-            raise ValueError(
-                "Entered `model_architecture` argument is not a standard available model_architecture. Valid options are ['fcn8s', 'fcn16s', 'fcn32s']"
-            )
-        elif not isinstance(self.backbone, str) and isinstance(
-            self.model_architecture, None
-        ):
-            raise ValueError(
-                "Using `model_architecture` is not allowed when supplying a custom backbone. Valid options are ['fcn8s', 'fcn16s', 'fcn32s']"
-            )
-        else:
-            if self.model_architecture is None:
-                self.model_architecture = "fcn8s"
-            if self.model_architecture == "fcn8s":
-                self.pool3 = tf.keras.Sequential(
-                    self.backbone_base_model.layers[
-                        : BACKBONE_CONFIG[self.backbone_name]["BLOCK3"]
-                    ]
-                )
-                self.pool4 = tf.keras.Sequential(
-                    self.backbone_base_model.layers[
-                        : BACKBONE_CONFIG[self.backbone_name]["BLOCK4"]
-                    ]
-                )
-                self.pool5 = tf.keras.Sequential(
-                    self.backbone_base_model.layers[
-                        : BACKBONE_CONFIG[self.backbone_name]["BLOCK5"]
-                    ]
-                )
-            elif self.model_architecture == "fcn16s":
-                self.pool4 = tf.keras.Sequential(
-                    self.backbone_base_model.layers[
-                        : BACKBONE_CONFIG[self.backbone_name]["BLOCK4"]
-                    ]
-                )
-                self.pool5 = tf.keras.Sequential(
-                    self.backbone_base_model.layers[
-                        : BACKBONE_CONFIG[self.backbone_name]["BLOCK5"]
-                    ]
-                )
-            elif self.model_architecture == "fcn32s":
-                self.pool5 = tf.keras.Sequential(
-                    self.backbone_base_model.layers[
-                        : BACKBONE_CONFIG[self.backbone_name]["BLOCK5"]
-                    ]
-                )
-
-        if self.model_architecture is not None:
-            units = [
-                i.units
-                for i in [
-                    self.backbone_base_model.layers[
-                        BACKBONE_CONFIG[self.backbone_name]["DENSE_START"]
-                    ]
-                ]
-            ]
-            self.dense_convs = [
-                tf.keras.layers.Conv2D(
-                    filters=i,
+            else:
+                backbone = CustomArchitectureBuilder(backbone)
+                classes_conv = tf.keras.layers.Conv2D(
+                    filters=classes,
                     kernel_size=(1, 1),
                     strides=(1, 1),
-                    activation="relu",
                     padding="same",
+                    activation="softmax",
                 )
-                for i in units
-            ]
-        else:
-            units = [
-                i.units
-                for i in self.backbone_base_model.layers
-                if isinstance(i, tf.keras.layers.Dense)
-            ]
-            if len(units) != 0:
-                self.dense_convs = []
-                for idx in len(units):
-                    curr_unit = units[idx]
-                    self.dense_convs.append(
-                        tf.keras.layers.Conv2D(
-                            filters=curr_unit,
-                            strides=(1, 1),
-                            activation="relu",
-                            padding="same",
-                        )
-                    )
-                    if idx != len(units) - 1:
-                        self.dense_convs.append(tf.keras.layers.Dropout(0.5))
-            output_shape = self.dense_convs[-1].compute_output_shape()
-            target_height_factor = self.height // output_shape[1]
-            target_width_factor = self.width // output_shape[2]
-            if output_shape[3] != self.num_classes:
-                self.dense_convs.append(
-                    tf.keras.layers.Conv2D(
-                        filters=self.num_classes,
-                        strides=(1, 1),
-                        activation="relu",
-                        padding="same",
-                    )
+
+                input_tensor = tf.keras.Input(input_shape)
+                if include_rescaling:
+                    x = Rescaling(scale=1.0 / 255)(input_tensor)
+                    x = backbone(x)
+                else:
+                    x = backbone(input_tensor)
+
+                output_shape = x.shape
+
+                target_height_factor = input_shape[0] // output_shape[1]
+                target_width_factor = input_shape[1] // output_shape[2]
+
+                upscale = tf.keras.layers.UpSampling2D(
+                    size=(target_height_factor, target_width_factor),
+                    data_format="channels_last",
+                    interpolation="bilinear",
                 )
-            self.output_layer = tf.keras.layers.UpSampling2D(
-                size=(target_height_factor, target_width_factor),
-                data_format="channels_last",
-                interpolation="bilinear",
-            )
 
-    def call(self, x):
-        if self.resize:
-            x = self.resizing_layer(x)
+                x = classes_conv(x)
+                output_tensor = upscale(x)
+                if return_mask:
+                    # Assumes channels_last
+                    output_tensor = tf.math.argmax(output_tensor, axis=3)
+                    output_tensor = tf.expand_dims(output_tensor, axis=3)
 
-        if self.model_architecture == "fcn8s":
-            pool3_output = self.pool3(x)
-            pool4_output = self.pool4(x)
-            pool5_output = self.pool5(x)
+                if return_dtype != output_tensor.dtype:
+                    output_tensor = tf.cast(output_tensor, return_dtype)
 
-            for layer in self.dense_convs:
-                pool5_output = layer(pool5_output)
-            pool5_output = tf.keras.layers.UpSampling2D(
-                size=(2, 2), data_format="channels_last", interpolation="bilinear"
-            )(pool5_output)
+                super().__init__(
+                    inputs={"input_tensor": input_tensor}, outputs=[output_tensor]
+                )
 
-            pool3_output = tf.keras.layers.Conv2D(
-                filters=pool5_output.shape[-1],
-                kernel_size=(1, 1),
-                padding="same",
-                strides=(1, 1),
-            )(pool3_output)
-            pool4_output = tf.keras.layers.Conv2D(
-                filters=pool5_output.shape[-1],
-                kernel_size=(1, 1),
-                padding="same",
-                strides=(1, 1),
-            )(pool4_output)
+                self.classes = classes
+                self.model_architecture = "custom"
+                self.include_rescaling = include_rescaling
+                self.upscale = upscale
+                self.backbone = backbone
+                self.classes_conv = classes_conv
+                self.target_height_factor = target_height_factor
+                self.target_width_factor = target_width_factor
+                self.return_dtype = return_dtype
+                self.return_mask = return_mask
+        elif isinstance(backbone, str):
+            if model_architecture not in ["fcn8s", "fcn16s", "fcn32s"]:
+                raise ValueError(
+                    "Invalid argument for parameter `model_architecture`. Accepted values are ['fcn8s', 'fcn16s', 'fcn32s']"
+                )
+            else:
+                input_tensor = tf.keras.Input(shape=input_shape)
+                output_tensor = VGGArchitectureBuilder(
+                    classes=classes,
+                    model_version=backbone,
+                    model_architecture=model_architecture,
+                    include_rescaling=include_rescaling,
+                    input_tensor=input_tensor,
+                    input_shape=input_shape,
+                    include_imagenet_weights=include_imagenet_weights,
+                )
+                if return_mask:
+                    # Assumes channels_last
+                    output_tensor = tf.math.argmax(output_tensor, axis=3)
+                    output_tensor = tf.expand_dims(output_tensor, axis=3)
 
-            intermediate_pool_output = tf.keras.layers.Add()(
-                [pool4_output, pool5_output]
-            )
-            intermediate_pool_output = tf.keras.layers.UpSampling2D(
-                size=(2, 2), data_format="channels_last", interpolation="bilinear"
-            )(intermediate_pool_output)
+                if return_dtype != output_tensor.dtype:
+                    output_tensor = tf.cast(output_tensor, return_dtype)
 
-            final_pool_output = tf.keras.layers.Add()(
-                [pool3_output, intermediate_pool_output]
-            )
-
-            output_layer = tf.keras.layers.UpSampling2D(
-                size=(8, 8), data_format="channels_last", interpolation="bilinear"
-            )
-            return output_layer(final_pool_output)
-
-        elif self.model_architecture == "fcn16s":
-            pool4_output = self.pool4(x)
-            pool5_output = self.pool5(x)
-
-            for layer in self.dense_convs:
-                pool5_output = layer(pool5_output)
-            pool5_output = tf.keras.layers.UpSampling2D(
-                size=(2, 2), data_format="channels_last", interpolation="bilinear"
-            )(pool5_output)
-
-            pool4_output = tf.keras.layers.Conv2D(
-                filters=pool5_output.shape[-1],
-                kernel_size=(1, 1),
-                padding="same",
-                strides=(1, 1),
-            )(pool4_output)
-
-            final_pool_output = tf.keras.layers.Add()(
-                [pool4_output, intermediate_pool_output]
-            )
-
-            output_layer = tf.keras.layers.UpSampling2D(
-                size=(16, 16), data_format="channels_last", interpolation="bilinear"
-            )
-            return output_layer(final_pool_output)
-
-        elif self.model_architecture == "fcn32s":
-            pool5_output = self.pool5(x)
-
-            for layer in self.dense_convs:
-                pool5_output = layer(pool5_output)
-            output_layer = tf.keras.layers.UpSampling2D(
-                size=(32, 32), data_format="channels_last", interpolation="bilinear"
-            )
-            return output_layer(pool5_output)
-
+                super().__init__(
+                    inputs={"input_tensor": input_tensor},
+                    outputs=[output_tensor],
+                )
         else:
-            x = self.backbone(x)
-            for layer in self.dense_convs:
-                x = layer(x)
-            return self.output_layer(x)
-
-    def train_step(self, data):
-        images, y_true, sample_weight = tf.keras.utils.unpack_x_y_sample_weight(data)
-        with tf.GradientTape() as tape:
-            y_pred = self.call(images)
-            total_loss = self.compute_loss(
-                x=images, y=y_true, y_pred=y_pred, sample_weight=sample_weight
+            raise ValueError(
+                """Invalid argument for parameter `backbone`. Accepted values are ['vgg16', 'vgg19'] or a `tf.keras.models.Model`
+                instance with only `tf.keras.layers.Conv2D`, 'tf.keras.layers.MaxPooling2D' or `tf.keras.layers.Dense` layers"""
             )
-        self.optimizer.minimize(total_loss, self.trainable_variables, tape=tape)
-        return self.compute_metrics(
-            x=images, y=y_true, y_pred=y_pred, sample_weight=sample_weight
-        )
 
     def get_config(self):
         config = {
             "classes": self.num_classes,
             "backbone": self.backbone,
             "model_architecture": self.model_architecture,
+            "return_dtype": self.return_dtype,
+            "return_mask": self.return_mask,
+            "include_rescaling": self.include_rescaling,
         }
         base_config = super().get_config()
         return dict(list(base_config.items()) + list(config.items()))
+
+
+def FCN8S(
+    classes,
+    include_rescaling,
+    backbone,
+    return_mask=False,
+    input_shape=(224, 224, 3),
+    return_dtype=tf.float32,
+    include_imagenet_weights=False,
+):
+    if not isinstance(backbone, str) or backbone not in ["vgg16", "vgg19"]:
+        raise ValueError(
+            "Invalid argument for parameter `backbone`. Accepted values are ['vgg16', 'vgg19']."
+        )
+    return FCN(
+        classes=classes,
+        input_shape=input_shape,
+        include_rescaling=include_rescaling,
+        backbone=backbone,
+        model_architecture="fcn8s",
+        return_mask=return_mask,
+        return_dtype=return_dtype,
+        include_imagenet_weights=include_imagenet_weights,
+    )
+
+
+def FCN16S(
+    classes,
+    include_rescaling,
+    backbone,
+    return_mask=False,
+    input_shape=(224, 224, 3),
+    return_dtype=tf.float32,
+    include_imagenet_weights=False,
+):
+    if not isinstance(backbone, str) or backbone not in ["vgg16", "vgg19"]:
+        raise ValueError(
+            "Invalid argument for parameter `backbone`. Accepted values are ['vgg16', 'vgg19']."
+        )
+    return FCN(
+        classes=classes,
+        input_shape=input_shape,
+        include_rescaling=include_rescaling,
+        backbone=backbone,
+        model_architecture="fcn16s",
+        return_mask=return_mask,
+        return_dtype=return_dtype,
+        include_imagenet_weights=include_imagenet_weights,
+    )
+
+
+def FCN32S(
+    classes,
+    include_rescaling,
+    backbone,
+    return_mask=False,
+    input_shape=(224, 224, 3),
+    return_dtype=tf.float32,
+    include_imagenet_weights=False,
+):
+    if not isinstance(backbone, str) or backbone not in ["vgg16", "vgg19"]:
+        raise ValueError(
+            "Invalid argument for parameter `backbone`. Accepted values are ['vgg16', 'vgg19']."
+        )
+    return FCN(
+        classes=classes,
+        input_shape=input_shape,
+        include_rescaling=include_rescaling,
+        backbone=backbone,
+        model_architecture="fcn32s",
+        return_mask=return_mask,
+        return_dtype=return_dtype,
+        include_imagenet_weights=include_imagenet_weights,
+    )
